magit-diff-update-hunk-region: New method for v25
The trick of using overlays containing a 1 pixel high newline to render
magit's region lines causes various glitches in the display
engine (overlays containing newlines tend to cause trouble).

When running in Emacs 25, instead render the top line as an :overline
face, and the bottom as an :underline on the first and last line of the
region, respectively.  While extending the lines past the text until the
edge of the screen does cause some complications (and messes with
temporary-goal-column movement), these are less serious than those
caused by the original method.